move MemoryAdapter from ts7200.rs -> memory module
